fix: 서버 응답 구조에 맞게 DTO 수정
- maxHeartRateConfig 타입 Int → Double (서버가 190.0으로 반환)
- HealthSummaryResponseDto에 summary nested 구조 반영

+++ Runnect-iOS/Runnect-iOS/Network/Dto/RunningDto/RequestDto/HealthDataSaveRequestDto.swift
@@ -19,7 +19,7 @@ struct HealthDataSaveRequestDto: Codable {
     let zone3Seconds: Int
     let zone4Seconds: Int
     let zone5Seconds: Int
-    let maxHeartRateConfig: Int?
+    let maxHeartRateConfig: Double?
     let heartRateSamples: [HeartRateSampleDto]?
 }
